+++ doc/changelog.rst
@@ -6,6 +6,9 @@ Release history
 
 This is a bugfix release
 
+* There were no 'classifiers' in the package metadata due to a bug
+  in setup.py.
+
 0.10.3
 ------