+++ doc/changelog.rst
@@ -25,6 +25,11 @@ Release history
 * Added ModuleGraph.getReferers(tonode). This methods yields the
   nodes that are referencing *tonode* (the reverse of getReferences)
 
+0.11.2
+------
+
+*
+
 0.11.1
 ------